+++ Android/OpenRegatta/src/com/openregatta/services/NMEADataFrame.java
@@ -23,27 +23,27 @@ public class NMEADataFrame {
 		/** 
 		 * WGS84 latitude degrees positive to the North
 		 */
-		public double LatitudeN = -1;
+		public float LatitudeN = -1;
 		/** 
 		 * WGS84 longitude degrees positive to the East
 		 */
-		public double LongitudeE = -1;
+		public float LongitudeE = -1;
 		/** 
 		 * Heading WGD84 terrestrial degrees
 		 */
-		public double HeadingT = -1;
+		public float HeadingT = -1;
 		/** 
 		 * Heading magnetic degrees
 		 */
-		public double HeadingM = -1;
+		public float HeadingM = -1;
 		/**
 		 * Speed over water in meters per second
 		 */
-		public double SpeedOverWater = -1;
+		public float SpeedOverWater = -1;
 		/**
 		 * Speed over ground in meters per second
 		 */
-		public double SpeedOverGround = -1;
+		public float SpeedOverGround = -1;
 	}
 	
 
@@ -55,28 +55,28 @@ public class NMEADataFrame {
 		 * Apparent wind angle degrees
 		 * 0 to 360 degrees with 0 as dead ahead increasing when rotating clockwise
 		 */
-		public double ApparentWindAngle = -1;
+		public float ApparentWindAngle = -1;
 		/** 
 		 * True wind angle in degrees
 		 * 0 to 360 degrees with 0 as dead ahead increasing when rotating clockwise
 		 */
-		public double TrueWindAngle = -1;
+		public float TrueWindAngle = -1;
 		/**
 		 * Apparent wind speed in meters per second
 		 */
-		public double ApparentWindSpeed = -1;
+		public float ApparentWindSpeed = -1;
 		/**
 		 * True wind direction WGS84 degrees terrestrial
 		 */
-		public double TrueWindDirectionT = -1;
+		public float TrueWindDirectionT = -1;
 		/** 
 		 * True wind direction magnetic
 		 */
-		public double TrueWindDirectionM = -1;
+		public float TrueWindDirectionM = -1;
 		/**
 		 * True wind speed in meters per second
 		 */
-		public double TrueWindSpeed = -1;
+		public float TrueWindSpeed = -1;
 	}
 	
 	public class Course
@@ -84,10 +84,10 @@ public class NMEADataFrame {
 		/**
 		 * Bearing to waypoint in degrees
 		 */
-		public double BearingToWaypoint = -1;
+		public float BearingToWaypoint = -1;
 		/**
 		 * Distance to waypoiny in meters
 		 */
-		public double DistanceToWaypoint = -1;
+		public float DistanceToWaypoint = -1;
 	}
 }
